requirements.txt
coverage
inquirerPy
fastapi
psycopg2-binary
pylint
pytest
python-dotenv
PyYAML
regex
requests
tabulate
uvicorn
